[NativeMenu] Move files to the `display` subfolder.
