+++ servers/register_server_types.cpp
@@ -57,11 +57,11 @@
 #include "camera/camera_feed.h"
 #include "camera_server.h"
 #include "debugger/servers_debugger.h"
+#include "display/native_menu.h"
 #include "display_server.h"
 #include "movie_writer/movie_writer.h"
 #include "movie_writer/movie_writer_mjpeg.h"
 #include "movie_writer/movie_writer_pngwav.h"
-#include "native_menu.h"
 #include "rendering/renderer_compositor.h"
 #include "rendering/renderer_rd/framebuffer_cache_rd.h"
 #include "rendering/renderer_rd/storage_rd/render_data_rd.h"
